Fix complilation error

+++ ex12/ft_print_memory.c
@@ -44,11 +44,12 @@ void	ft_write_safe_char(char *c)
 		write(1, &".", 1);
 }
 
+
 void	ft_print_memory_at(void *start_addr, unsigned int size, char *curr_addr)
 {
 	int	index;
 
-	ft_write_hex((unsigned int)curr_addr, 16, MAX_PAGE_SIZE);
+	ft_write_hex((unsigned long)curr_addr, 16, MAX_PAGE_SIZE);
 	write(1, &": ", 2);
 	index = 0;
 	while (index++ < 16)
